% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_p_adjust.R
\name{.log_BY}
\alias{.log_BY}
\title{Multiple Test Correction: BY}
\usage{
.log_BY(p, n, lp, base)
}
\arguments{
\item{p}{x}

\item{n}{x}

\item{lp}{x}

\item{base}{x}
}
\value{
a vector of adjusted p-values in log space
}
\description{
Multiple Test Correction: BY
}
