% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_p_adjust.R
\name{.log_factory}
\alias{.log_factory}
\title{log function factory helper}
\usage{
.log_factory(base = exp(1))
}
\arguments{
\item{base}{numeric value to set the base of the log function to. default: exp(1)}
}
\value{
function
}
\description{
Helper function meant to be used internally.
Returns a function that represents \code{log(x, base = base)}
}
\examples{
custom_log10 <- .log_factory(10)
custom_log10(10)

}
