% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_p_adjust.R
\docType{data}
\name{logp.adjust.methods}
\alias{logp.adjust.methods}
\title{Methods supported by logp.adjust}
\format{
A character vector.
}
\usage{
logp.adjust.methods
}
\description{
Character vector of method names accepted by \code{logp.adjust()}.
These are derived from p.adjust.methods and should match unless a new method has been added.
}
\examples{
logp.adjust.methods
}
\seealso{
\link{p.adjust.methods}
}
\keyword{datasets}
