% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_p_adjust.R
\name{logp.adjust}
\alias{logp.adjust}
\title{Adjust log P-values for Multiple Comparisons}
\usage{
logp.adjust(p, method = logp.adjust.methods, n = length(p), base = exp(1))
}
\arguments{
\item{p}{vector of p-values in log space}

\item{method}{correction method to use}

\item{n}{number of comparisons, must be greater than or equal to length(p).}

\item{base}{the log base to perform operations in. Should match the base of the p-values. defaults to natural log.}
}
\value{
vector of adjusted log p-values
}
\description{
Adjust log P-values for Multiple Comparisons
}
